% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_hist.R
\name{box_hist}
\alias{box_hist}
\title{Create histograms with horizontal boxplots above}
\usage{
box_hist(.data, mapping, binwidth = 1, title = NULL, subtitle = NULL,
  x_label = NULL, y_label = NULL, x_limits = NULL, y_limits = NULL,
  fill = "#69727a", color = "#414649", box_color = NULL,
  y_format = NULL, line_intercept = NULL)
}
\arguments{
\item{line_intercept}{}
}
\description{
A similar visualization can be created with
}
