% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_coef.R
\name{my_coef}
\alias{my_coef}
\title{Title}
\usage{
my_coef(.data, x_mapping = "estimate", y_mapping = "term",
  exclude_intercept = FALSE, title = "Coefficients plot",
  subtitle = "with 95\% confidence interval", x_label = "",
  y_label = "", x_limits = NULL, y_limits = NULL, fill = "#69727a",
  color = "#414649", breaks = NULL, x_digits = 1)
}
\arguments{
\item{breaks}{}
}
\description{
Title
}
