% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_hist.R
\name{my_hist}
\alias{my_hist}
\title{Define function to create histograms}
\usage{
my_hist(.data, mapping, group_mapping = NULL, binwidth = 1,
  title = NULL, subtitle = NULL, x_label = NULL, y_label = NULL,
  x_limits = NULL, y_limits = NULL, fill = "#69727a",
  color = "#414649", y_format = NULL, line_intercept = NULL)
}
\arguments{
\item{line_intercept}{}
}
\description{
Define function to create histograms
}
