% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_waffle.R
\name{my_waffle}
\alias{my_waffle}
\title{Simple wrapper around waffle()}
\usage{
my_waffle(.data, title = NULL,
  x_lab = "(One Square Represents 10,000 Visual Fields)",
  colors = c("#414649", "#69727a", "#afafaf"), rows = 10, size = 2,
  legend_pos = "top", font_family = "Arial")
}
\arguments{
\item{font_family}{}
}
\description{
Simple wrapper around waffle()
}
