% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip_empty.R
\name{strip_empty}
\alias{strip_empty}
\title{Define internal function to strip empty grobs from facet plot}
\usage{
strip_empty(p)
}
\arguments{
\item{p}{}
}
\description{
Define internal function to strip empty grobs from facet plot
}
