% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_heatmap.R
\name{vf_heatmap}
\alias{vf_heatmap}
\title{Visual Fields Heatmap}
\usage{
vf_heatmap(.data, mapping, title = NULL, subtitle = NULL,
  fill_limits = NULL, low_color = "black", high_color = "white",
  legend_position = "none", decimals = 2)
}
\arguments{
\item{decimals}{}
}
\description{
Visual Fields Heatmap
}
